refactor(guard): bound native receipt components
