+++ ci/native_runtime/probe_native_default_auto.py
@@ -13,6 +13,7 @@ import urllib.error
 import urllib.parse
 import urllib.request
 from pathlib import Path
+from typing import cast
 
 import codex_plugin_scanner
 from codex_plugin_scanner.guard.adapters.codex_daemon_hook_transport import (
@@ -23,6 +24,9 @@ from codex_plugin_scanner.guard.daemon.server import GuardDaemonServer
 from codex_plugin_scanner.guard.native_policy_test_support import native_policy_snapshot
 from codex_plugin_scanner.guard.native_resident_client import native_resident_client_failure_code
 from codex_plugin_scanner.guard.native_runtime import (
+    NativeRuntimeCapabilities,
+    NativeRuntimeIdentity,
+    NativeRuntimeStatus,
     native_mode,
     native_runtime_health,
     native_runtime_status,
@@ -147,6 +151,83 @@ def _installed_hook_request(
         ) from error
 
 
+def _exercise_installed_routes(
+    daemon: GuardDaemonServer,
+    guard_home: Path,
+    workspace: Path,
+    routes: dict[str, dict[str, str]],
+    route_receipts: list[dict[str, str]],
+    reason_codes: dict[str, int],
+) -> None:
+    for harness, route in sorted(routes.items()):
+        events: list[tuple[str, dict[str, object]]] = []
+        if route["pre_tool_use"].startswith("installed_"):
+            events.append(
+                (
+                    "PreToolUse",
+                    {
+                        "hook_event_name": "PreToolUse",
+                        "tool_name": "Bash",
+                        "tool_input": {"command": "printf guard"},
+                    },
+                )
+            )
+        if route["post_tool_use"].startswith("installed_"):
+            events.append(
+                (
+                    "PostToolUse",
+                    {
+                        "hook_event_name": "PostToolUse",
+                        "tool_name": "Read",
+                        "tool_response": [{"type": "text", "text": "guard baseline\n"}],
+                    },
+                )
+            )
+        for event, payload in events:
+            response_payload = _installed_hook_request(daemon, guard_home, workspace, harness, event, payload)
+            if response_payload is None:
+                raise RuntimeError(f"empty response for {harness} {event}")
+            reason = response_payload.get("reason_code")
+            if isinstance(reason, str):
+                reason_codes[reason] = reason_codes.get(reason, 0) + 1
+            route_receipts.append({"harness": harness, "event": event, "route": "native_resident"})
+
+
+def _exercise_mode_invariants(
+    daemon: GuardDaemonServer,
+    guard_home: Path,
+    workspace: Path,
+) -> dict[str, dict[str, object]]:
+    mode_invariants: dict[str, dict[str, object]] = {}
+    try:
+        for mode in ("off", "shadow"):
+            os.environ["HOL_GUARD_NATIVE"] = mode
+            response = _installed_hook_request(
+                daemon,
+                guard_home,
+                workspace,
+                "claude-code",
+                "PostToolUse",
+                {
+                    "hook_event_name": "PostToolUse",
+                    "tool_name": "Read",
+                    "tool_response": [{"type": "text", "text": "mode invariant\n"}],
+                },
+            )
+            if not isinstance(response, dict):
+                raise RuntimeError(f"native_default_auto_probe_failed: invalid mode response: {response}")
+            _require(response.get("decision") == "block", {"mode": mode, "response": response})
+            mode_invariants[mode] = {
+                "decision": response.get("decision"),
+                "reason_code": response.get("reason_code"),
+                "python_oracle": daemon._server.hook_worker.test_oracle is not None,
+            }
+            _require(mode_invariants[mode]["python_oracle"] is False, mode_invariants[mode])
+    finally:
+        os.environ.pop("HOL_GUARD_NATIVE", None)
+    return mode_invariants
+
+
 def _installed_hook_corpus(root: Path) -> dict[str, object]:
     guard_home = root / "hook-home"
     workspace = root / "hook-workspace"
@@ -167,7 +248,6 @@ def _installed_hook_corpus(root: Path) -> dict[str, object]:
     if callable(register_workspace):
         _ = register_workspace(workspace)
     reason_codes: dict[str, int] = {}
-
     route_receipts: list[dict[str, str]] = []
     routes = _ownership_routes()
     daemon.start()
@@ -186,79 +266,18 @@ def _installed_hook_corpus(root: Path) -> dict[str, object]:
                 "policy_ready": prepared_policy is not None,
             },
         )
-        for harness, route in sorted(routes.items()):
-            events: list[tuple[str, dict[str, object]]] = []
-            if route["pre_tool_use"].startswith("installed_"):
-                events.append(
-                    (
-                        "PreToolUse",
-                        {
-                            "hook_event_name": "PreToolUse",
-                            "tool_name": "Bash",
-                            "tool_input": {"command": "printf guard"},
-                        },
-                    )
-                )
-            if route["post_tool_use"].startswith("installed_"):
-                events.append(
-                    (
-                        "PostToolUse",
-                        {
-                            "hook_event_name": "PostToolUse",
-                            "tool_name": "Read",
-                            "tool_response": [{"type": "text", "text": "guard baseline\n"}],
-                        },
-                    )
-                )
-            for event, payload in events:
-                response_payload = _installed_hook_request(
-                    daemon,
-                    guard_home,
-                    workspace,
-                    harness,
-                    event,
-                    payload,
-                )
-                if response_payload is None:
-                    raise RuntimeError(f"empty response for {harness} {event}")
-                reason = response_payload.get("reason_code")
-                if isinstance(reason, str):
-                    reason_codes[reason] = reason_codes.get(reason, 0) + 1
-                route_receipts.append({"harness": harness, "event": event, "route": "native_resident"})
-        # Native edge requests execute in the resident HookWorker and do not
-        # traverse the compatibility subprocess runner. Read its bounded
-        # route receipt instead of attributing native traffic to that idle
-        # Python-only worker pool.
+        _exercise_installed_routes(daemon, guard_home, workspace, routes, route_receipts, reason_codes)
         worker_stats = daemon._server.hook_worker.metrics.snapshot()
         writer = daemon._server.runtime_hook_evidence_writer
-        for mode in ("off", "shadow"):
-            os.environ["HOL_GUARD_NATIVE"] = mode
-            response = _installed_hook_request(
-                daemon,
-                guard_home,
-                workspace,
-                "claude-code",
-                "PostToolUse",
-                {
-                    "hook_event_name": "PostToolUse",
-                    "tool_name": "Read",
-                    "tool_response": [{"type": "text", "text": "mode invariant\n"}],
-                },
-            )
-            _require(isinstance(response, dict), {"mode": mode, "response": response})
-            _require(response.get("decision") == "block", {"mode": mode, "response": response})
-            mode_invariants[mode] = {
-                "decision": response.get("decision"),
-                "reason_code": response.get("reason_code"),
-                "python_oracle": daemon._server.hook_worker.test_oracle is not None,
-            }
-            _require(mode_invariants[mode]["python_oracle"] is False, mode_invariants[mode])
+        mode_invariants = _exercise_mode_invariants(daemon, guard_home, workspace)
     finally:
-        os.environ.pop("HOL_GUARD_NATIVE", None)
         daemon.stop()
 
     expected = len(route_receipts)
-    observed_routes = worker_stats["routes"]
+    observed_routes_raw = worker_stats["routes"]
+    if not isinstance(observed_routes_raw, dict):
+        raise RuntimeError(f"native_default_auto_probe_failed: invalid route metrics: {worker_stats}")
+    observed_routes = cast(dict[str, int], observed_routes_raw)
     evidence_stats = writer.stats()
     _require(expected > 0, "installed hook corpus is empty")
     _require(expected == 21, {"expected": expected, "routes": routes})
@@ -287,7 +306,7 @@ def _installed_hook_corpus(root: Path) -> dict[str, object]:
     }
 
 
-def main(*, json_path: Path | None = None) -> int:
+def _require_clean_probe_environment() -> None:
     for environment_name in (
         "HOL_GUARD_NATIVE",
         "HOL_GUARD_NATIVE_BINARY",
@@ -299,33 +318,24 @@ def main(*, json_path: Path | None = None) -> int:
         _require(environment_name not in os.environ, f"{environment_name} must be unset")
     _require(native_mode() == "auto", f"unexpected native mode: {native_mode()}")
     _require(hook_fast_path_enabled(), "unset fast-path configuration must be enabled")
-
     os.environ["HOL_GUARD_NATIVE"] = "invalid"
     try:
         _require(native_mode() == "auto", "invalid native mode must resolve to auto")
     finally:
-        del os.environ["HOL_GUARD_NATIVE"]
+        os.environ.pop("HOL_GUARD_NATIVE", None)
 
-    package_path = Path(codex_plugin_scanner.__file__).resolve()
-    source_package = (Path.cwd() / "src" / "codex_plugin_scanner").resolve()
-    _require(
-        not package_path.is_relative_to(source_package),
-        f"probe imported source tree package: {package_path}",
-    )
 
+def _probe_native_identity() -> tuple[NativeRuntimeStatus, NativeRuntimeIdentity, NativeRuntimeCapabilities]:
     status = native_runtime_status()
     _require(status.mode == "auto", status)
     _require(status.available and status.compatible, status)
     _require(status.reason == "native_ready", status)
-    identity = status.identity
-    if identity is None:
+    if status.identity is None or status.capabilities is None:
         raise RuntimeError(f"native_default_auto_probe_failed: {status}")
-    capabilities = status.capabilities
-    if capabilities is None:
-        raise RuntimeError(f"native_default_auto_probe_failed: {status}")
+    return status, status.identity, status.capabilities
 
-    # An invalid binary override must not redirect automatic production
-    # selection away from the version-bound bundled resident runtime.
+
+def _assert_binary_override_ignored(identity: NativeRuntimeIdentity) -> None:
     os.environ["HOL_GUARD_NATIVE_BINARY"] = "/definitely-missing/hol-guard-runtime"
     try:
         overridden = native_runtime_status()
@@ -337,45 +347,51 @@ def main(*, json_path: Path | None = None) -> int:
             {"selected": overridden.identity, "expected": identity.path},
         )
     finally:
-        del os.environ["HOL_GUARD_NATIVE_BINARY"]
+        os.environ.pop("HOL_GUARD_NATIVE_BINARY", None)
 
+
+def _run_native_smoke(root: Path) -> None:
+    with native_policy_snapshot(root / "guard-home") as snapshot:
+        clean = review_post_tool_native(
+            _request(root, "const value = 1;\n", "default-auto-clean"),
+            observe_mode=False,
+            policy_snapshot=snapshot,
+        )
+        if clean is None:
+            raise RuntimeError(
+                "native_default_auto_probe_failed: clean response missing: "
+                f"{native_resident_client_failure_code()}"
+            )
+        _require(clean.decision == "allow", clean)
+        secret = review_post_tool_native(
+            _request(root, _synthetic_github_token(), "default-auto-secret"),
+            observe_mode=False,
+            policy_snapshot=snapshot,
+        )
+        if secret is None:
+            raise RuntimeError("native_default_auto_probe_failed: secret response missing")
+        _require(secret.decision == "deny", secret)
+        _require(secret.reason_code == "output_secret_match", secret)
+
+
+def _run_temporary_probe(identity: NativeRuntimeIdentity) -> dict[str, object]:
     with tempfile.TemporaryDirectory(prefix="hg-auto-", dir=_short_temp_parent()) as temporary:
         root = Path(temporary)
         try:
-            with native_policy_snapshot(root / "guard-home") as snapshot:
-                clean = review_post_tool_native(
-                    _request(root, "const value = 1;\n", "default-auto-clean"),
-                    observe_mode=False,
-                    policy_snapshot=snapshot,
-                )
-                if clean is None:
-                    raise RuntimeError(
-                        "native_default_auto_probe_failed: clean response missing: "
-                        f"{native_resident_client_failure_code()}"
-                    )
-                _require(clean.decision == "allow", clean)
-
-                secret = review_post_tool_native(
-                    _request(root, _synthetic_github_token(), "default-auto-secret"),
-                    observe_mode=False,
-                    policy_snapshot=snapshot,
-                )
-                if secret is None:
-                    raise RuntimeError("native_default_auto_probe_failed: secret response missing")
-                _require(secret.decision == "deny", secret)
-                _require(secret.reason_code == "output_secret_match", secret)
-
+            _run_native_smoke(root)
             health = native_runtime_health(root / "guard-home")
             _require(health.state == "healthy", health)
             _require(health.reason == "native_ready", health)
             _require(health.resident_failures == 0, health)
             _require(health.oneshot_failures == 0, health)
             _require(len(_native_state_files(root / "guard-home")) == 1, "native generation was not reused")
-            installed_corpus = _installed_hook_corpus(root)
+            return _installed_hook_corpus(root)
         finally:
             for guard_home in (root / "guard-home", root / "hook-home"):
                 _stop_native_runtime(identity.path, guard_home)
 
+
+def _assert_native_disabled_mode() -> None:
     os.environ["HOL_GUARD_NATIVE"] = "off"
     try:
         _require(native_mode() == "off", f"unexpected native mode: {native_mode()}")
@@ -383,9 +399,15 @@ def main(*, json_path: Path | None = None) -> int:
         _require(disabled.mode == "off", disabled)
         _require(disabled.reason == "native_disabled", disabled)
     finally:
-        del os.environ["HOL_GUARD_NATIVE"]
+        os.environ.pop("HOL_GUARD_NATIVE", None)
 
-    receipt = {
+
+def _build_probe_receipt(
+    status: NativeRuntimeStatus,
+    capabilities: NativeRuntimeCapabilities,
+    installed_corpus: dict[str, object],
+) -> dict[str, object]:
+    return {
         "schema": "hol-guard.native-default-installed-receipt.v1",
         "corpus_scope": "normalized_daemon_ingress",
         "default_mode": "auto",
@@ -404,6 +426,22 @@ def main(*, json_path: Path | None = None) -> int:
         "receipt_metrics": installed_corpus["receipt_metrics"],
         "mode_invariants": installed_corpus["mode_invariants"],
     }
+
+
+def main(*, json_path: Path | None = None) -> int:
+    _require_clean_probe_environment()
+    package_path = Path(codex_plugin_scanner.__file__).resolve()
+    source_package = (Path.cwd() / "src" / "codex_plugin_scanner").resolve()
+    _require(
+        not package_path.is_relative_to(source_package),
+        f"probe imported source tree package: {package_path}",
+    )
+
+    status, identity, capabilities = _probe_native_identity()
+    _assert_binary_override_ignored(identity)
+    installed_corpus = _run_temporary_probe(identity)
+    _assert_native_disabled_mode()
+    receipt = _build_probe_receipt(status, capabilities, installed_corpus)
     rendered = json.dumps(receipt, sort_keys=True)
     if json_path is not None:
         json_path.write_text(rendered + "\n", encoding="utf-8")
