+++ ci/test-suite-ratchet-baseline.json
@@ -1,5 +1,5 @@
 {
   "schema_version": 1,
-  "maximum_collected_cases": 16621,
-  "maximum_test_source_lines": 360317
+  "maximum_collected_cases": 16625,
+  "maximum_test_source_lines": 360450
 }
