+++ src/codex_plugin_scanner/guard/adapters/cursor_hooks.py
@@ -59,12 +59,7 @@ _INHERIT_ENV_KEYS = (
     "CURSOR_TRANSCRIPT_PATH",
     "HOL_GUARD_NATIVE",
     "HOL_GUARD_NATIVE_BINARY",
-    "HOL_GUARD_SRC",
-    # These markers are forwarded only when an explicit test harness requests
-    # the differential oracle; production installations leave them unset.
-    "HOL_GUARD_TEST_MODE",
-    "HOL_GUARD_PYTHON_ORACLE",
-    "HOL_GUARD_NATIVE_DIAGNOSTIC",
+    "HOL_GUARD_SRC", "HOL_GUARD_TEST_MODE", "HOL_GUARD_PYTHON_ORACLE", "HOL_GUARD_NATIVE_DIAGNOSTIC",
 )
