+++ src/codex_plugin_scanner/guard/daemon/server.py
@@ -6194,9 +6194,7 @@ class _GuardDaemonHandler(BaseHTTPRequestHandler):
         if review.payload is not None and time.monotonic() < process_deadline:
             if review.receipt is not None:
                 with suppress(Exception):
-                    _ = daemon_server.runtime_hook_evidence_writer.submit_native_decision_receipt(
-                        receipt=review.receipt,
-                    )
+                    _ = daemon_server.runtime_hook_evidence_writer.submit_native_decision_receipt(review.receipt)
             self._write_json(review.payload)
             return
         reason_code = (
