+++ src/codex_plugin_scanner/guard/store_connection_schema.py
@@ -11,7 +11,7 @@ from datetime import datetime, timezone
 from typing import ClassVar
 from uuid import uuid4
 
-from . import store_review_event_outbox_schema
+from . import store_native_decision_receipts, store_review_event_outbox_schema
 from .mcp.policy_store import ensure_mcp_policy_request_schema
 from .sqlite_profile import (
     SQLiteMigrationGateReport,
@@ -35,11 +35,6 @@ from .store_command_activity_schema import ensure_command_activity_schema
 from .store_command_shadow_schema import ensure_command_shadow_schema
 from .store_extension_control_authority_schema import ensure_extension_control_authority_schema
 from .store_local_cli_schema import ensure_local_cli_schema
-from .store_native_decision_receipts import (
-    NATIVE_DECISION_RECEIPT_MIGRATION_VERSION,
-    native_decision_receipt_index_statements,
-    native_decision_receipt_schema_statement,
-)
 from .store_resume import ensure_resume_schema
 from .store_review_event_outbox_schema import ensure_review_event_outbox_schema
 from .store_secret_policy_integrity import _POLICY_INTEGRITY_LOOKUP_UNSET
@@ -167,9 +162,8 @@ _REQUIRED_SCHEMA_MIGRATION_VERSIONS = (  # Keep retired-index databases on the p
     WORKFLOW_CAPABILITY_RECEIPT_EVENT_INDEX_MIGRATION_VERSION,
     WATCH_ONLY_APPROVAL_MIGRATION_VERSION,
     store_review_event_outbox_schema.REVIEW_EVENT_OUTBOX_MIGRATION_VERSION,
-    NATIVE_DECISION_RECEIPT_MIGRATION_VERSION,
+    *store_native_decision_receipts.native_decision_receipt_migration_versions(),
 )
-
 
 @dataclass
 class _SchemaInitializationState:
@@ -982,8 +976,7 @@ class StoreConnectionSchemaMixin:
             supply_chain_bundle_schema_statement(),
             supply_chain_eval_cache_schema_statement(),
             threat_intel_bundle_schema_statement(),
-            threat_intel_matches_schema_statement(),
-            native_decision_receipt_schema_statement(),
+            *store_native_decision_receipts.native_decision_receipt_schema_statements(threat_intel_matches_schema_statement()),
         )
         with self._connect() as connection:
             if initialize_incremental_vacuum:
@@ -994,8 +987,6 @@ class StoreConnectionSchemaMixin:
             # readers exhaust the bounded hook write deadline.
             self._enable_wal_mode(connection)
             for statement in statements:
-                connection.execute(statement)
-            for statement in native_decision_receipt_index_statements():
                 connection.execute(statement)
             ensure_resume_schema(connection)
             ensure_command_activity_schema(connection, applied_at=_now())
@@ -1122,10 +1113,6 @@ class StoreConnectionSchemaMixin:
             self._record_schema_version(
                 connection,
                 version=STORAGE_QUERY_INDEX_MIGRATION_VERSION,
-            )
-            self._record_schema_version(
-                connection,
-                version=NATIVE_DECISION_RECEIPT_MIGRATION_VERSION,
             )
             if not self._schema_version_applied(connection, version=2):
                 self._record_schema_version(connection, version=2)
