+++ tests/test_guard_hook_process_runner.py
@@ -684,12 +684,8 @@ def test_prewarmed_runner_scans_post_tool_output_in_isolated_worker(tmp_path: Pa
 
     assert result.reason_code is None
     assert result.payload is not None
-    # This worker-pool test runs under the explicit, test-only Python oracle
-    # fixture. Native semantic authority is covered by the installed/runtime
-    # suites; the pool assertion here verifies the compatibility oracle result
-    # survives the isolated evaluator boundary.
-    assert result.payload["recorded"] is True
-    assert result.payload["policy_action"] == "warn"
+    # Explicit test oracle; native terminal paths are covered by runtime suites.
+    assert result.payload["recorded"] is True and result.payload["policy_action"] == "warn"
     assert runner.stats()["workers"] == 0
 
 
@@ -729,8 +725,7 @@ def test_idempotent_review_retries_once_after_worker_death(tmp_path: Path) -> No
 
     assert result.reason_code is None
     assert result.payload is not None
-    assert result.payload["recorded"] is True
-    assert result.payload["policy_action"] == "warn"
+    assert result.payload["recorded"] is True and result.payload["policy_action"] == "warn"
 
 
 def test_worker_retry_withdraws_scheduler_capacity_before_reusing_slot(
